+++ lib/subsocia_common.mli
@@ -53,6 +53,7 @@ module Type : sig
 
   val equal : 'a t -> 'b t -> bool
   val compare : 'a t -> 'b t -> int
+  val pp : Format.formatter -> 'a t -> unit
 end
 
 module Value : sig
@@ -109,8 +110,8 @@ module Values : sig
   val compare : 'a t -> 'a t -> int
 
   val coerce : 'a Type.t -> ex -> 'a t
-    [@@ocaml.deprecated "Use coerce_exn."] [@@ocaml.warning "-3"]
-  val coerce_exn : 'a Type.t -> any -> 'a t
+    [@@ocaml.deprecated "Use coerce_any."] [@@ocaml.warning "-3"]
+  val coerce_any : 'a Type.t -> any -> 'a t option
 
   val to_json : 'a Type.t ->
     'a t ->